Refactored MainFragment out of activity and renamed
