+++ app/src/main/res/layout-land/phrase_input_fragment.xml
