+++ app/src/main/res/layout/phrase_input_fragment.xml
